requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-autocomplete-light==3.3.2
django-ckeditor==5.2.2
django-croppie==1.3
django-phonenumber-field==2.1.0
idna==2.8

oauthlib==2.1.0
olefile==0.46
phonenumbers==8.10.2
Pillow==5.3.0
psycopg2==2.7.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
-e git://github.com/mariocesar/sorl-thumbnail.git@7e61948e7a987714097ba50a54120f584db9c843#egg=sorl_thumbnail
urllib3==1.24.1
vine==1.1.4
